requirements.txt
# MCP Framework
fastmcp>=2.0.0

# HTTP Client
httpx>=0.27.0

# Environment management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
respx>=0.21.0
faker>=22.0.0
